+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.25] - 2024-12-26
+
 ### Added
 
+- Add customer credit information on totalizers
 - Node and graphql setup for saved carts
+- Applying a saved cart on checkout
 
 ### Changed